Project structure refactoring
